# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  code: string;
  message: string;
  requestId?: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  advertiser: 'advertiser',
  creator: 'creator',
  reviewer: 'reviewer',
  accountant: 'accountant',
  admin: 'admin',
  super_admin: 'super_admin',
  ai_operator: 'ai_operator',
} as const;

export interface User {
  id: string;
  clerkUserId: string;
  email?: string | null;
  displayName?: string | null;
  role: UserRole;
  isActive: boolean;
  createdAt: string;
}

export type CampaignStatus = typeof CampaignStatus[keyof typeof CampaignStatus];


export const CampaignStatus = {
  draft: 'draft',
  under_review: 'under_review',
  approved: 'approved',
  funding: 'funding',
  active: 'active',
  paused: 'paused',
  completed: 'completed',
  settlement: 'settlement',
  closed: 'closed',
  archived: 'archived',
} as const;

export type CampaignTarget = { [key: string]: unknown };

export type CampaignRequirements = { [key: string]: unknown };

export interface Campaign {
  id: string;
  organizationId: string;
  createdByUserId: string;
  name: string;
  brief?: string | null;
  status: CampaignStatus;
  /** @minimum 0 */
  budgetMinor: number;
  currency: string;
  /** @minimum 0 */
  cpvMinor: number;
  /** @minimum 0 */
  targetViews: number;
  target: CampaignTarget;
  requirements: CampaignRequirements;
  platforms: string[];
  startsOn?: string | null;
  endsOn?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateCampaignRequestTarget = { [key: string]: unknown };

export type CreateCampaignRequestRequirements = { [key: string]: unknown };

export interface CreateCampaignRequest {
  /** @minLength 1 */
  name: string;
  brief?: string | null;
  /** @minimum 0 */
  budgetMinor: number;
  /**
     * @minLength 3
     * @maxLength 3
     */
  currency: string;
  /** @minimum 0 */
  cpvMinor: number;
  /** @minimum 0 */
  targetViews: number;
  target?: CreateCampaignRequestTarget;
  requirements?: CreateCampaignRequestRequirements;
  platforms?: string[];
  startsOn?: string | null;
  endsOn?: string | null;
}

/**
 * Draft campaign fields to update
 */
export type UpdateCampaignRequest = CreateCampaignRequest;

export interface CampaignList {
  items: Campaign[];
  nextCursor?: string | null;
}

export interface CreateApplicationRequest {
  pitch?: string | null;
}

export type ApplicationStatus = typeof ApplicationStatus[keyof typeof ApplicationStatus];


export const ApplicationStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  withdrawn: 'withdrawn',
} as const;

export interface Application {
  id: string;
  campaignId: string;
  creatorId: string;
  status: ApplicationStatus;
  pitch?: string | null;
  createdAt: string;
}

export type ContentSubmissionStatus = typeof ContentSubmissionStatus[keyof typeof ContentSubmissionStatus];


export const ContentSubmissionStatus = {
  submitted: 'submitted',
  in_review: 'in_review',
  approved: 'approved',
  rejected: 'rejected',
  needs_changes: 'needs_changes',
  escalated: 'escalated',
} as const;

export interface ContentSubmission {
  id: string;
  campaignId: string;
  creatorId: string;
  applicationId?: string | null;
  postUrl: string;
  caption?: string | null;
  hashtags: string[];
  platform: string;
  status: ContentSubmissionStatus;
  submittedAt: string;
  updatedAt: string;
}

export type ReviewDecisionRequestDecision = typeof ReviewDecisionRequestDecision[keyof typeof ReviewDecisionRequestDecision];


export const ReviewDecisionRequestDecision = {
  approved: 'approved',
  rejected: 'rejected',
  needs_changes: 'needs_changes',
  escalated: 'escalated',
} as const;

export type ReviewDecisionRequestChecklist = {[key: string]: boolean};

export interface ReviewDecisionRequest {
  decision: ReviewDecisionRequestDecision;
  checklist?: ReviewDecisionRequestChecklist;
  reason?: string | null;
  note?: string | null;
}

export type ReviewMode = typeof ReviewMode[keyof typeof ReviewMode];


export const ReviewMode = {
  human: 'human',
  ai: 'ai',
} as const;

export interface ReviewModeRequest {
  mode: ReviewMode;
}

export interface ReviewModeResponse {
  mode: ReviewMode;
  changedAt: string;
}

/**
 * Authentication required
 */
export type UnauthorizedResponse = Error;

/**
 * Permission denied
 */
export type ForbiddenResponse = Error;

/**
 * Invalid request
 */
export type ValidationErrorResponse = Error;

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

export type ListCampaignsParams = {
status?: CampaignStatus;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};