# lib/api-zod/src/generated/types/campaignStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type CampaignStatus = typeof CampaignStatus[keyof typeof CampaignStatus];


export const CampaignStatus = {
  draft: 'draft',
  under_review: 'under_review',
  approved: 'approved',
  funding: 'funding',
  active: 'active',
  paused: 'paused',
  completed: 'completed',
  settlement: 'settlement',
  closed: 'closed',
  archived: 'archived',
} as const;