# lib/api-zod/src/generated/types/createCampaignRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCampaignRequestRequirements } from './createCampaignRequestRequirements';
import type { CreateCampaignRequestTarget } from './createCampaignRequestTarget';

export interface CreateCampaignRequest {
  /** @minLength 1 */
  name: string;
  brief?: string | null;
  /** @minimum 0 */
  budgetMinor: number;
  /**
     * @minLength 3
     * @maxLength 3
     */
  currency: string;
  /** @minimum 0 */
  cpvMinor: number;
  /** @minimum 0 */
  targetViews: number;
  target?: CreateCampaignRequestTarget;
  requirements?: CreateCampaignRequestRequirements;
  platforms?: string[];
  startsOn?: Date | null;
  endsOn?: Date | null;
}