# lib/api-zod/src/generated/types/reviewDecisionRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewDecisionRequestChecklist } from './reviewDecisionRequestChecklist';
import type { ReviewDecisionRequestDecision } from './reviewDecisionRequestDecision';

export interface ReviewDecisionRequest {
  decision: ReviewDecisionRequestDecision;
  checklist?: ReviewDecisionRequestChecklist;
  reason?: string | null;
  note?: string | null;
}