# lib/api-zod/src/generated/types/reviewModeResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewMode } from './reviewMode';

export interface ReviewModeResponse {
  mode: ReviewMode;
  changedAt: Date;
}