# lib/api-zod/src/generated/types/updateCampaignRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCampaignRequest } from './createCampaignRequest';

/**
 * Draft campaign fields to update
 */
export type UpdateCampaignRequest = CreateCampaignRequest;