# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: string;
  clerkUserId: string;
  email?: string | null;
  displayName?: string | null;
  role: UserRole;
  isActive: boolean;
  createdAt: Date;
}