# lib/api-zod/src/generated/types/userRole.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  advertiser: 'advertiser',
  creator: 'creator',
  reviewer: 'reviewer',
  accountant: 'accountant',
  admin: 'admin',
  super_admin: 'super_admin',
  ai_operator: 'ai_operator',
} as const;